refactor!: merge pre_commit_hooks into repoma
