+++ setup.cfg
@@ -81,11 +81,11 @@ dev =
 
 [options.entry_points]
 console_scripts =
-    check-dev-files = repoma.pre_commit_hooks.check_dev_files:main
-    fix-first-nbcell = repoma.pre_commit_hooks.fix_first_nbcell:main
-    fix-nbformat-version = repoma.pre_commit_hooks.fix_nbformat_version:main
-    format-setup-cfg = repoma.pre_commit_hooks.format_setup_cfg:main
-    pin-nb-requirements = repoma.pre_commit_hooks.pin_nb_requirements:main
+    check-dev-files = repoma.check_dev_files:main
+    fix-first-nbcell = repoma.fix_first_nbcell:main
+    fix-nbformat-version = repoma.fix_nbformat_version:main
+    format-setup-cfg = repoma.format_setup_cfg:main
+    pin-nb-requirements = repoma.pin_nb_requirements:main
 
 [options.packages.find]
 where = src
