+++ src/repoma/check_dev_files/check_labels.py
@@ -9,7 +9,7 @@ import pathlib
 from typing import TYPE_CHECKING
 
 from repoma._utilities import CONFIG_PATH
-from repoma.pre_commit_hooks.errors import PrecommitError
+from repoma.errors import PrecommitError
 
 if TYPE_CHECKING:
     from typing import List
