+++ src/repoma/check_dev_files/flake8.py
@@ -13,7 +13,7 @@ from repoma._utilities import (
     open_config,
     open_setup_cfg,
 )
-from repoma.pre_commit_hooks.errors import PrecommitError
+from repoma.errors import PrecommitError
 
 if TYPE_CHECKING:
     from configparser import ConfigParser
