--- src/repoma/pre_commit_hooks/__init__.py
@@ -1,1 +1,0 @@
-"""Python scripts with entry points for the pre-commit hooks."""
