+++ tests/pre_commit_hooks/test_github_templates.py
@@ -1,7 +1,5 @@
 from repoma._utilities import REPOMA_DIR
-from repoma.pre_commit_hooks.check_dev_files.github_templates import (
-    _list_template_files,
-)
+from repoma.check_dev_files.github_templates import _list_template_files
 
 
 def test_list_issue_templates():
